thermal: of: add support name based cooling binding
There is possibility that thermal zone and related sensor driver
is registered through the DT but the cooling device is through
board files. On this case, cooling device binding with thermal
zone is done using the type of the cooling device and the cooling 
device map.

This way, it will be possible to support the thermal zone from DT and
cooling device from non-DT.

Change-Id: Ia394e9deb62496d256e86fcbd8b3459a9297cf2c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/419504