ARM: tegra: clock: Add PTO counters in debugfs
Added support for reading the PTO counter in debugfs. Clock that can
be monitored get a new debugfs entry 'pto_rate' which will trigger a
PTO measurement, and report the result provided the clock is enabled.

This commit is based on the original work by
Peter De Schrijver <pdeschrijver@nvidia.com>

Change-Id: I549263595cbde49cee4adbebd29513e087790951
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/501449
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>