iommu/arm-smmu: caps: add IOMMU_CAP_INTR_REMAP capability
MSIs are just seen as bog standard memory writes by the ARM SMMU, so
they can be translated (and isolated) in the same way.

This patch adds the IOMMU_CAP_INTR_REMAP capability to the ARM SMMU
driver and reworks our capabaility code so that we don't assume the
caps are organised as bits in a bitmask (since this isn't the intention).

Change-Id: Ic0e4763bb25d378931f40cb75708ed5b20e0f785
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>