mmu_notifier: Add invalidate_range_free_pages() notifier v2
This notifier closes an important gap in the current
invalidate_range_start()/end() notifiers. The _start() part
is called when all pages are still mapped while the _end()
notifier is called when all pages are potentially unmapped
and already freed.

This does not allow to manage external (non-CPU) hardware
TLBs with MMU-notifiers because there is no way to prevent
that hardware will esablish new TLB entries between the
calls of these two functions. But this is a requirement to
the subsytem that implements these existing notifiers.

To allow managing external TLBs the MMU-notifiers need to
catch the moment when pages are unmapped but not yet freed.
This new notifier catches that moment and notifies the
interested subsytem when pages that were unmapped are about
to be freed. The new notifier will only be called between
invalidate_range_start()/end().

Changed since v1:
  - Rebase on top of lastest linux next.
  - Expect vma and range to be provided to callback.

Signed-off-by: Joerg Roedel <joro@8bytes.org>
Signed-off-by: Jérôme Glisse <jglisse@redhat.com>
(cherry picked from commit 086155fd0e89e6c6a211cb114841e29076e633cb)
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>

Conflicts:
	include/linux/mmu_notifier.h

Bug 200074285

Change-Id: Ic63c83173aa50f4439cd02d72dd0590eed5e86bb
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/678277
Reviewed-by: Hiroshi Doyu <hdoyu@nvidia.com>