iommu/arm-smmu: fix capability checking prior to device attach
If somebody attempts to check the capability of an IOMMU domain prior to
device attach, then we'll try to dereference a NULL SMMU pointer through
the SMMU domain (since we can't determine the actual SMMU instance until
we have a device attached).

This patch fixes the capability check so that non-global features are
reported as being absent when no device is attached to the domain.

Change-Id: I896fd4809eaa8c70a00317a70c6d5a730c5252eb
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>