arm64: t210: ensure wake status is correctly captured
This is to ensure the wake status is correctly captured
and held correctly for all the wake.  This is necessitated
due to a bug in the logic which had made the E_INPUT
control for specific pin i.e. GPIO_PA6 to Logic0 because
of which wake event status didn.t get latched in some
corner cases.

SW WAR:
1. Enable WAK_DET_EN while entering LP0.
2. During warm boot exit code- program E_INPUT of gpio_pa6 to one;
3. Restore whole pinmux programming and clear PARK bits.
4. De-assert WAKE_DET_EN.

Step 2 is implemented in warmboot code.

Bug 1518986

Change-Id: I99cd06ba208020cd4f19caf678cee2285df73e9e
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/440783
Reviewed-by: Sachin Nikam <snikam@nvidia.com>