cpuidle: tegra210: Add cluster dependent states
Added support for cluster dependent target_residency and exit_latency
field for each state. This is required because the fast and the slow
cluster could have vastly different energy costs for the same idle
states.

Also added debugfs support to enable/disable states per-cluster. The
disabled flag in the cpuidle framework disables the state for both
clusters. If the menu governor asks for a state that has been disabled
at the cluster level, the driver will look through and find the next
possible deepest state that has been enabled both at the cpuidle
framework level and the per-cluster level in the driver.

Change-Id: I9acf4eeec87da4581eec4b753997f715d1c9f763
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Signed-off-by: Prashant Gaikwad <pgaikwad@nvidia.com>
Reviewed-on: http://git-master/r/421895
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>