mmc: tegra: sleep/awake support during lp0 based on plat-data.
On automotive board configurations, PMU shuts off both the powers
going to the eMMC device (VCC and VCCQ). Due to this,
response of CMD5(Awake) has timeout error during resume,
if the device is put in SLEEP.

This change makes the decision to sleep the card or shutdown
the card, based on input from platform-data.

Bug 1360277
Bug 1448682

Change-Id: I45fe6e4bed69948696f06763f7d1bc62dd7f5682
Signed-off-by: Seshagir.H <sholi@nvidia.com>
Reviewed-on: http://git-master/r/456661
Reviewed-by: Jinyoung Park <jinyoungp@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>