arm: tegra: 1859: Disable eMMC sleep functionality
On some board configurations,PMU shuts off both the powers
going to the eMMC device (VCC and VCCQ). Due to this, response
of CMD5(Awake) has timeout error during resume, if the device
is put in SLEEP.

This change makes the decision to sleep the card or shutdown
the card, based on input from sdhci instance's DT data.

Bug 1360277
Bug 1448682

Change-Id: Ie83b9e235e29fbe24bccce5ce8e32555892b258d
Signed-off-by: Seshagir.H <sholi@nvidia.com>
Reviewed-on: http://git-master/r/456663
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Tested-by: Sandeep Trasi <strasi@nvidia.com>