media: tegra_v4l2_camera: fix format support
This driver does not demosaic, so when we are working with a bayer sensor,
we can only report the various bayer formats.  Similarly, we only support
rgba for test pattern generator right now, so only report rgba if we are in
test pattern mode.

When in test pattern mode, don't ever query the sensor subdev, since it's
not involved at all.

Also, made some miscellaneous cleanup to some debug prints, and the
conditionals around how CSI gets programmed.

Bug 1560636

Change-Id: I9022ec6a489b20d0847dd9d2a79af2307b7a4af1
Signed-off-by: Andrew Chew <achew@nvidia.com>
Signed-off-by: Bryan Wu <pengw@nvidia.com>
Reviewed-on: http://git-master/r/489066
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: David Wang (SW-TEGRA) <davidw@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>