hid: atv-remote: change to new mSBC packet protocol
New protocol supports encoding keycodes in the first
mSBC packet so that we can support reporting keys
pressed/released without an additional packet in the
connection interval.  We have to take the packet out
and send it to the report that is for inputs.

Change-Id: Ice7d04201d9575b19e42329e7fe8a799c5bc985d
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Xia Yang <xiay@nvidia.com>
Reviewed-on: http://git-master/r/670842
(cherry picked from commit 3ff3c5ac72031ec713cd619de24a789944e04329)
Reviewed-on: http://git-master/r/601316
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>