platform: nvadsp: Fix DRAM logger issue on restart
Logging stops working once ADSP restarts. This is since the write pointer on
ADSP resets while the read pointer is waiting for an update(as in waiting on
EOT). Moving the read pointer to the intial position and marking it EOT so that
it waits till ADSP boot up and updates the DRAM logger memory.

Bug 200031688

Change-Id: I365f180c51662a70cb95dafb96dc98c7949415bd
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: http://git-master/r/493976
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Nitin Kumbhar <nkumbhar@nvidia.com>