ARM: DT: vcm30t124: panel color depth by EDID
Panel color depth (bit per pixel) is taken from the EDID for most panels.
But, the depth setting at the platform data default_out has the highest
priority and will override the EDID if it is not zero.
This DT change sets the default_out->depth to zero to make the EDID color
depth info effective.

Change-Id: I01d362e60a376256035f9a6f175113131a9d0796
Signed-off-by: Sungwook Kim <sungwookk@nvidia.com>
Reviewed-on: http://git-master/r/494962
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>