iommu/tegra: smmu: retrieve AS props from domain
The address space properties were attached to each device,
this does not confirm to upstream iommus bindings and this
approach is not extendible to devices which still rely on fixup
tables.

Retrieve the asprop info from DT during probe and store it in
a list. Since, the number of domains are just 5, traversing the list,
although O(n) does not slowdown the initialization much.

Also, add support to retrieve swgids from iommus= as per upstream bindings.
At this point, we still support the downstream way of handling things but
it will be removed a little bit later.

Bug 200032393
Bug 200020468

Change-Id: I3c75207110e6aa79296d1fd7a45046f065e6a759
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/495653
Reviewed-by: Hiroshi Doyu <hdoyu@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User