security: tlk_driver: add session tracking
Previously, if a client application exits w/o
closing any open session(s) the session and
its resources would be lost.

With this change we now track open sessions so
that any that remain open when tlk_device_release
is called can be properly closed.

This allows memory buffer tracking to be simplified
as well since the various lists can be tracked
per-session instead of globally as before.

Bug 1551921

Change-Id: I134ab7f2a963dc9afba41f5411187d61dc3b6539
Signed-off-by: Scott Long <scottl@nvidia.com>