arm64: dts: Remove bootargs from chosen node
The shim layer and cboot shall append to the kernel command line
present in the /chosen/bootargs.

Right now, cboot overwrites to this property anyway, to avoid conflicts.
Hence, we should be good to remove chosen/bootargs. This is done to be able to
add a property to command line easily, through the top level DTS, if needed,
later on.

Bug 1552234

Change-Id: Ia57a41ed1ea68e45bc3e0fea15b93e07b5c06b0c
Signed-off-by: Jay Bhukhanwala <jbhukhanwala@nvidia.com>
Reviewed-on: http://git-master/r/499591
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>