asoc: codec: audience: Don't clear cachedcmd_list
Don't clear cachedcmd_list from Algorithm kcontrol

The cachedcmd_list now holds the refcount information for each MUX which
is set twice. If it is cleared from the Algorithm kcontrol when it is
set to Off, it would erase the refcount information as well which is
required to update the power state of DAPM widgets when refcount reaches
to zero.

Added an extra check for skipping the power state update for DAPM
widget. In case of dragon board XML, if the playback is running and the
capture is started via Sound Recorder app, the MUXes are not set twice
as they are already set. But the disable sequence of playback gets
executed. In this case, the refcount of all the MUXes would be
decremented in disable sequence, and that would break the DAPM route
even if the TX (capture) stream is still active. The extra check of
active streams before updating the power state of DAPM widget will keep
the DAPM route as it is.

bug 200037978

Change-Id: I397a22e59d95fb4474d93aaf8e5a7def0d24c222
Signed-off-by: Dara Ramesh <dramesh@nvidia.com>
Reviewed-on: http://git-master/r/500174
Reviewed-by: Viraj Karandikar <vkarandikar@nvidia.com>
Tested-by: Viraj Karandikar <vkarandikar@nvidia.com>
Reviewed-by: Vinod Subbarayalu <vsubbarayalu@nvidia.com>
Tested-by: Vinod Subbarayalu <vsubbarayalu@nvidia.com>